+++ seed/python-sdk/auth-environment-variables/src/seed/core/client_wrapper.py
@@ -6,7 +6,8 @@ import httpx
 
 
 class BaseClientWrapper:
-    def __init__(self, *, api_key: str, base_url: str):
+    def __init__(self, *, x_another_header: str, api_key: str, base_url: str):
+        self._x_another_header = x_another_header
         self.api_key = api_key
         self._base_url = base_url
 
@@ -16,6 +17,7 @@ class BaseClientWrapper:
             "X-Fern-SDK-Name": "seed",
             "X-Fern-SDK-Version": "0.0.0",
         }
+        headers["X-Another-Header"] = self._x_another_header
         headers["X-FERN-API-KEY"] = self.api_key
         return headers
 
@@ -24,12 +26,12 @@ class BaseClientWrapper:
 
 
 class SyncClientWrapper(BaseClientWrapper):
-    def __init__(self, *, api_key: str, base_url: str, httpx_client: httpx.Client):
-        super().__init__(api_key=api_key, base_url=base_url)
+    def __init__(self, *, x_another_header: str, api_key: str, base_url: str, httpx_client: httpx.Client):
+        super().__init__(x_another_header=x_another_header, api_key=api_key, base_url=base_url)
         self.httpx_client = httpx_client
 
 
 class AsyncClientWrapper(BaseClientWrapper):
-    def __init__(self, *, api_key: str, base_url: str, httpx_client: httpx.AsyncClient):
-        super().__init__(api_key=api_key, base_url=base_url)
+    def __init__(self, *, x_another_header: str, api_key: str, base_url: str, httpx_client: httpx.AsyncClient):
+        super().__init__(x_another_header=x_another_header, api_key=api_key, base_url=base_url)
         self.httpx_client = httpx_client